requirements_test.txt
-e .

codecov==2.0.9
flake8==3.0.4
freezegun==0.3.8
pytest-capturelog==0.7
pytest-cov==2.3.1
pytest-django==3.0.0
pytest-sugar==0.8.0
pytest==3.0.2
requests_mock==1.1.0
setuptools>=38.6.0,<39.0.0
twine>=1.11.0,<2.0.0
wheel>=0.31.0,<1.0.0
